data_collection_and_annotation/requirements.txt
torch
transformers
opencv-python
pandas
numpy
Pillow